+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `AcceptedAbstractSyntaxes: []string{"*"}` accepts any peer-proposed Abstract Syntax (generic Storage SCP)
+
 ## [0.9.0] - 2026-07-14
 
 ### Added